requirements.txt
opencv-python>=4.8.0
numpy>=1.24.0
pyzbar>=0.1.9
pillow>=10.0.0
pyobjc-framework-AVFoundation>=10.0  # For iOS camera access
pyobjc-framework-Cocoa>=10.0
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.7.0  # For code formatting
mypy>=1.5.0   # For type checking